VIP-63 Live Preview
It's not 100% live, you need to click a button to
reload the preview with the new config. I did have
it live reacting but re-initializing the VIT app
on every keypress of a title or custom link, for
instance, resulted in a broken VIT sometimes. So
this way is better.